requirements.txt
google-generativeai
streamlit
pillow
streamlit-calendar